Crate civo_rs

This crate provides a client for interacting with the Civo API.

Modules

client

Module containing the client definition and related helper function

disk_image

Module for managing disk images, useful when working with instances

errors

Module containing errors that could be returned by the API

instance

Module containing instance definitions and related functions

instance_size

Module for retreiving instance sizes on civo

kubernetes

Module for managing Kubernetes clusters

network

Module containing network related structs and functions

objectstore

Module for managing Objectstores

region

Module containing Civo region related objects

ssh_key

Module for managing ssh keys

utils

Module cotaining misc helper fucntions including the heroku style name generator